fine tuned upvote container
