+++ src/components/upvote/index.js
@@ -1,5 +1,4 @@
-import UpvoteView from './view/upvoteView';
 import Upvote from './container/upvoteContainer';
 
-export { UpvoteView, Upvote };
+export { Upvote };
 export default Upvote;
